Moved ThriftClientHandler, AssignmentHandler and UnloadTabletHandler into separate classes (#1581)
