+++ server/tserver/src/main/java/org/apache/accumulo/tserver/AssignmentHandler.java
@@ -1,0 +1,243 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.accumulo.tserver;
+
+import static org.apache.accumulo.server.problems.ProblemType.TABLET_LOAD;
+
+import java.util.Arrays;
+import java.util.Set;
+import java.util.SortedSet;
+import java.util.TimerTask;
+import java.util.TreeSet;
+
+import org.apache.accumulo.core.data.TableId;
+import org.apache.accumulo.core.dataImpl.KeyExtent;
+import org.apache.accumulo.core.master.thrift.TabletLoadState;
+import org.apache.accumulo.core.metadata.schema.TabletMetadata;
+import org.apache.accumulo.core.util.Daemon;
+import org.apache.accumulo.fate.util.LoggingRunnable;
+import org.apache.accumulo.server.master.state.Assignment;
+import org.apache.accumulo.server.master.state.TabletStateStore;
+import org.apache.accumulo.server.problems.ProblemReport;
+import org.apache.accumulo.server.problems.ProblemReports;
+import org.apache.accumulo.server.util.MasterMetadataUtil;
+import org.apache.accumulo.server.util.time.SimpleTimer;
+import org.apache.accumulo.tserver.TabletServerResourceManager.TabletResourceManager;
+import org.apache.accumulo.tserver.mastermessage.TabletStatusMessage;
+import org.apache.accumulo.tserver.tablet.Tablet;
+import org.apache.accumulo.tserver.tablet.TabletData;
+import org.apache.hadoop.io.Text;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class AssignmentHandler implements Runnable {
+  private static final Logger log = LoggerFactory.getLogger(AssignmentHandler.class);
+  private final KeyExtent extent;
+  private final int retryAttempt;
+  private final SortedSet<KeyExtent> unopenedTablets;
+  private final SortedSet<KeyExtent> openingTablets;
+  private final OnlineTablets onlineTablets;
+  private final TabletServer server;
+
+  public AssignmentHandler(TabletServer server, KeyExtent extent) {
+    this(server, extent, 0);
+  }
+
+  public AssignmentHandler(TabletServer server, KeyExtent extent, int retryAttempt) {
+    this.server = server;
+    this.onlineTablets = server.getOnlineTabletsRaw();
+    this.unopenedTablets = server.getUnopenedTablets();
+    this.openingTablets = server.getOpeningTablets();
+    this.extent = extent;
+    this.retryAttempt = retryAttempt;
+  }
+
+  @Override
+  public void run() {
+    synchronized (unopenedTablets) {
+      synchronized (openingTablets) {
+        synchronized (onlineTablets) {
+          // nothing should be moving between sets, do a sanity
+          // check
+          Set<KeyExtent> unopenedOverlapping = KeyExtent.findOverlapping(extent, unopenedTablets);
+          Set<KeyExtent> openingOverlapping = KeyExtent.findOverlapping(extent, openingTablets);
+          Set<KeyExtent> onlineOverlapping =
+              KeyExtent.findOverlapping(extent, onlineTablets.snapshot());
+
+          if (openingOverlapping.contains(extent) || onlineOverlapping.contains(extent)) {
+            return;
+          }
+
+          if (!unopenedOverlapping.contains(extent)) {
+            log.info("assignment {} no longer in the unopened set", extent);
+            return;
+          }
+
+          if (unopenedOverlapping.size() != 1 || openingOverlapping.size() > 0
+              || onlineOverlapping.size() > 0) {
+            throw new IllegalStateException(
+                "overlaps assigned " + extent + " " + !unopenedTablets.contains(extent) + " "
+                    + unopenedOverlapping + " " + openingOverlapping + " " + onlineOverlapping);
+          }
+        }
+
+        unopenedTablets.remove(extent);
+        openingTablets.add(extent);
+      }
+    }
+
+    // check Metadata table before accepting assignment
+    Text locationToOpen = null;
+    TabletMetadata tabletMetadata = null;
+    boolean canLoad = false;
+    try {
+      tabletMetadata = server.getContext().getAmple().readTablet(extent);
+
+      canLoad = TabletServer.checkTabletMetadata(extent, server.getTabletSession(), tabletMetadata);
+
+      if (canLoad && tabletMetadata.sawOldPrevEndRow()) {
+        KeyExtent fixedExtent =
+            MasterMetadataUtil.fixSplit(server.getContext(), tabletMetadata, server.getLock());
+
+        synchronized (openingTablets) {
+          openingTablets.remove(extent);
+          openingTablets.notifyAll();
+          // it expected that the new extent will overlap the old one... if it does not, it
+          // should not be added to unopenedTablets
+          if (!KeyExtent.findOverlapping(extent, new TreeSet<>(Arrays.asList(fixedExtent)))
+              .contains(fixedExtent)) {
+            throw new IllegalStateException(
+                "Fixed split does not overlap " + extent + " " + fixedExtent);
+          }
+          unopenedTablets.add(fixedExtent);
+        }
+        // split was rolled back... try again
+        new AssignmentHandler(server, fixedExtent).run();
+        return;
+
+      }
+    } catch (Exception e) {
+      synchronized (openingTablets) {
+        openingTablets.remove(extent);
+        openingTablets.notifyAll();
+      }
+      log.warn("Failed to verify tablet " + extent, e);
+      server.enqueueMasterMessage(new TabletStatusMessage(TabletLoadState.LOAD_FAILURE, extent));
+      throw new RuntimeException(e);
+    }
+
+    if (!canLoad) {
+      log.debug("Reporting tablet {} assignment failure: unable to verify Tablet Information",
+          extent);
+      synchronized (openingTablets) {
+        openingTablets.remove(extent);
+        openingTablets.notifyAll();
+      }
+      server.enqueueMasterMessage(new TabletStatusMessage(TabletLoadState.LOAD_FAILURE, extent));
+      return;
+    }
+
+    Tablet tablet = null;
+    boolean successful = false;
+
+    try {
+      server.acquireRecoveryMemory(extent);
+
+      TabletResourceManager trm = server.getResourceManager().createTabletResourceManager(extent,
+          server.getTableConfiguration(extent));
+      TabletData data = new TabletData(tabletMetadata);
+
+      tablet = new Tablet(server, extent, trm, data);
+      // If a minor compaction starts after a tablet opens, this indicates a log recovery
+      // occurred. This recovered data must be minor compacted.
+      // There are three reasons to wait for this minor compaction to finish before placing the
+      // tablet in online tablets.
+      //
+      // 1) The log recovery code does not handle data written to the tablet on multiple tablet
+      // servers.
+      // 2) The log recovery code does not block if memory is full. Therefore recovering lots of
+      // tablets that use a lot of memory could run out of memory.
+      // 3) The minor compaction finish event did not make it to the logs (the file will be in
+      // metadata, preventing replay of compacted data)... but do not
+      // want a majc to wipe the file out from metadata and then have another process failure...
+      // this could cause duplicate data to replay.
+      if (tablet.getNumEntriesInMemory() > 0
+          && !tablet.minorCompactNow(MinorCompactionReason.RECOVERY)) {
+        throw new RuntimeException("Minor compaction after recovery fails for " + extent);
+      }
+      Assignment assignment = new Assignment(extent, server.getTabletSession());
+      TabletStateStore.setLocation(server.getContext(), assignment);
+
+      synchronized (openingTablets) {
+        synchronized (onlineTablets) {
+          openingTablets.remove(extent);
+          onlineTablets.put(extent, tablet);
+          openingTablets.notifyAll();
+          server.getRecentlyUnloadedCache().remove(tablet.getExtent());
+        }
+      }
+      tablet = null; // release this reference
+      successful = true;
+    } catch (Throwable e) {
+      log.warn("exception trying to assign tablet {} {}", extent, locationToOpen, e);
+
+      if (e.getMessage() != null) {
+        log.warn("{}", e.getMessage());
+      }
+
+      TableId tableId = extent.getTableId();
+      ProblemReports.getInstance(server.getContext()).report(new ProblemReport(tableId, TABLET_LOAD,
+          extent.getUUID().toString(), server.getClientAddressString(), e));
+    } finally {
+      server.releaseRecoveryMemory(extent);
+    }
+
+    if (!successful) {
+      synchronized (unopenedTablets) {
+        synchronized (openingTablets) {
+          openingTablets.remove(extent);
+          unopenedTablets.add(extent);
+          openingTablets.notifyAll();
+        }
+      }
+      log.warn("failed to open tablet {} reporting failure to master", extent);
+      server.enqueueMasterMessage(new TabletStatusMessage(TabletLoadState.LOAD_FAILURE, extent));
+      long reschedule = Math.min((1L << Math.min(32, retryAttempt)) * 1000, 10 * 60 * 1000L);
+      log.warn(String.format("rescheduling tablet load in %.2f seconds", reschedule / 1000.));
+      SimpleTimer.getInstance(server.getConfiguration()).schedule(new TimerTask() {
+        @Override
+        public void run() {
+          log.info("adding tablet {} back to the assignment pool (retry {})", extent, retryAttempt);
+          AssignmentHandler handler = new AssignmentHandler(server, extent, retryAttempt + 1);
+          if (extent.isMeta()) {
+            if (extent.isRootTablet()) {
+              new Daemon(new LoggingRunnable(log, handler), "Root tablet assignment retry").start();
+            } else {
+              server.getResourceManager().addMetaDataAssignment(extent, log, handler);
+            }
+          } else {
+            server.getResourceManager().addAssignment(extent, log, handler);
+          }
+        }
+      }, reschedule);
+    } else {
+      server.enqueueMasterMessage(new TabletStatusMessage(TabletLoadState.LOADED, extent));
+    }
+  }
+}
